requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
databases
pymysql
passlib[bcrypt]
python-jose[cryptography]
python-multipart